Animated sun and moon
Animated the sun and moon, due to the day/night cycle not being 50/50
I had to do some calculations to calculate the position of the celestial object
Resizing the window throws the calculation off, but this can be explained
due to 1. clouds in day 2. iridium flare in the night :-D

Signed-off-by: Simon Egli <simon.egli@bbv.ch>